Add a TritonGPU LocalBarrierOp
Membar only tracks accesses to shared memory so all it needs is a fence
to the shared memory address space and a thread barrier.
Before this change a GPU BarrierOp was emitted which implies much
stronger guarantees on syncronization of all address spaces and not just
shared memory.

The new LocalBarrierOp is targeting only accesses between shared
memory and registers to allow for better code generation.

Added lowering for this new op:
- Common code lowers it to the current GPU BarrierOp
- AMD will lower to lgkmcnt(0) + s_barrier